requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
tensorflow>=2.15.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.28.0
openai>=1.0.0
